Refactor bulk export dependencies and remove unnecessary exclusions
