+++ java/bulk-export/bulk-export-core/pom.xml
@@ -31,15 +31,6 @@
             <artifactId>core</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
-        <dependency>
-            <groupId>sleeper</groupId>
-            <artifactId>common-task</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>software.amazon.awssdk</groupId>
-            <artifactId>ecs</artifactId>
-        </dependency>
 
         <!-- Test dependencies -->
         <dependency>
