+++ java/bulk-export/bulk-export-runner/pom.xml
@@ -51,6 +51,15 @@
             <artifactId>bulk-export-core</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
+        <dependency>
+            <groupId>sleeper</groupId>
+            <artifactId>common-task</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>software.amazon.awssdk</groupId>
+            <artifactId>ecs</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
