+++ java/bulk-export/bulk-export-runner/src/main/java/sleeper/bulkexport/runner/RunLeafPartitionBulkExportTasks.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.bulkexport.core.recordretrieval;
+package sleeper.bulkexport.runner;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
