+++ java/bulk-export/bulk-export-runner/src/main/java/sleeper/bulkexport/runner/RunSqsLeafPartitionBulkExportLambda.java
@@ -26,7 +26,6 @@ import org.slf4j.LoggerFactory;
 import software.amazon.awssdk.services.ecs.EcsClient;
 
 import sleeper.bulkexport.core.model.BulkExportLeafPartitionQuerySerDe;
-import sleeper.bulkexport.core.recordretrieval.RunLeafPartitionBulkExportTasks;
 import sleeper.configuration.properties.S3InstanceProperties;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.task.common.QueueMessageCount;
